+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## v 0.0.6
+
+- Color change according to the percentage of checked checkboxes
+- Mode for show both percentage and number of checked checkboxes
+- Fixed bug showing `NaN%` when there are no checkboxes in the document
+
 ## v 0.0.5
 
 - Updated `README.md` to include a link to the `CHANGELOG.md` file.